Add feature to generate an image CSV file.
