+++ tfrecorder/utils_test.py
@@ -16,6 +16,7 @@
 
 """Tests `utils.py`."""
 
+import csv
 import functools
 import os
 import tempfile
@@ -28,10 +29,10 @@ import tensorflow as tf
 
 from tfrecorder import beam_image
 from tfrecorder import constants
-from tfrecorder import utils
-from tfrecorder import test_utils
-from tfrecorder import input_schema
 from tfrecorder import dataset_loader
+from tfrecorder import input_schema
+from tfrecorder import test_utils
+from tfrecorder import utils
 
 
 # pylint: disable=protected-access
@@ -89,8 +90,8 @@ class CheckTFRecordsTest(unittest.TestCase):
 
       # Check output CSV
       actual_df = pd.read_csv(actual_csv)
-      expected_df = pd.DataFrame(self.data)
-      pdt.assert_frame_equal(actual_df, expected_df)
+      exp_df = pd.DataFrame(self.data)
+      pdt.assert_frame_equal(actual_df, exp_df)
 
       # Check output images
       actual_image_files = [
@@ -134,3 +135,85 @@ class CopyLogTest(unittest.TestCase):
       infile = os.path.join(tmpdirname, 'foo.txt')
       with self.assertRaises(FileNotFoundError):
         utils.copy_logfile_to_gcs(infile, tmpdirname)
+
+
+class PathSplitTest(unittest.TestCase):
+  """Tests `_path_split`."""
+
+  def test_local_and_gcs(self):
+    """Tests both local and GCS paths."""
+
+    filename = 'image_file.jpg'
+    dirpaths = ['/path/to/image/dir/', 'gs://path/to/image/dir/']
+    for dir_ in dirpaths:
+      filepath = os.path.join(dir_, filename)
+      act_dirpath, act_filename = utils._path_split(filepath)
+      self.assertEqual(act_dirpath, dir_.rsplit('/', 1)[0])
+      self.assertEqual(act_filename, filename)
+
+
+class ReadImageDirectoryTest(unittest.TestCase):
+  """Tests `read_image_directory`."""
+
+  def setUp(self):
+    self.image_data = test_utils.get_test_df()
+    self.tempfiles = []
+    self.tempdir = None
+    self.schema = input_schema.Schema(
+        input_schema.IMAGE_CSV_SCHEMA.input_schema_map)
+
+  def tearDown(self):
+    for fp in self.tempfiles:
+      fp.close()
+    self.tempdir.cleanup()
+
+  def test_normal(self):
+    """Tests conversion of expected directory structure on local machine."""
+
+    g = self.image_data.groupby([self.schema.split_key, self.schema.label_key])
+
+    self.tempdir = tempfile.TemporaryDirectory()
+    rows = []
+    for (split, label), indices in g.groups.items():
+      dir_ = os.path.join(self.tempdir.name, split, label)
+      os.makedirs(dir_)
+      for f in list(self.image_data.loc[indices, self.schema.image_uri_key]):
+        _, name = os.path.split(f)
+        fp = tempfile.NamedTemporaryFile(
+            dir=dir_, suffix='.jpg', prefix=name)
+        self.tempfiles.append(fp)
+        rows.append([split, fp.name, label])
+
+    columns = list(input_schema.IMAGE_CSV_SCHEMA.get_input_keys())
+    actual = utils.read_image_directory(self.tempdir.name)
+    actual.sort_values(by=columns, inplace=True)
+    actual.reset_index(drop=True, inplace=True)
+    expected = pd.DataFrame(rows, columns=columns)
+    expected.sort_values(by=columns, inplace=True)
+    expected.reset_index(drop=True, inplace=True)
+    pd.testing.assert_frame_equal(actual, expected)
+
+
+class CreateImageCSVTest(unittest.TestCase):
+  """Tests `create_image_csv`."""
+
+  @mock.patch.object(utils, 'read_image_directory', autospec=True)
+  def test_normal(self, mock_fn):
+    """Tests normal case."""
+
+    exp_df = test_utils.get_test_df()
+    mock_fn.return_value = exp_df
+    image_dir = 'path/to/image/dir'
+    with tempfile.NamedTemporaryFile(
+        mode='w+', prefix='image', suffix='.csv') as f:
+      utils.create_image_csv(image_dir, f.name)
+
+      # CSV should not have an index column
+      reader = csv.reader(f, delimiter=',')
+      row = next(reader)
+      exp_columns = exp_df.columns
+      self.assertEqual(len(row), len(exp_columns))
+
+      # CSV should match input data
+      actual = pd.read_csv(f.name, names=exp_columns)
+      pd.testing.assert_frame_equal(actual, exp_df)
